Refactor: modularized api services
