--- api/src/controllers/jobsControllers/jobs.ts
@@ -1,3 +1,0 @@
-import { query } from "express";
-
-
