+++ api/src/services/planService/PlanServiceAPI.ts
@@ -1,0 +1,8 @@
+import { PrismaClient } from "@prisma/client";
+import dotenv from "dotenv";
+dotenv.config();
+
+export const prisma = new PrismaClient();
+export const baseUrl = process.env.API_MP_URL;
+
+
